DOMCore/tests/submissions/Ms2ger/historical.html

<!DOCTYPE html>
<title>Historical DOM features must be removed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
function isInterfaceNuked(name) {
  test(function() { assert_equals(window[name], undefined) })
}
var nukedInterfaces = [
  "CDATASection",
  "DOMConfiguration",
  "DOMError",
  "DOMErrorHandler",
  "DOMImplementationList",
  "DOMImplementationSource",
  "DOMLocator",
  "DOMObject",
  "DOMUserData",
  "Entity",
  "EntityReference",
  "NamedNodeMap",
  "NameList",
  "Notation",
  "TypeInfo",
  "UserDataHandler"
]
nukedInterfaces.forEach(isInterfaceNuked)

function isNukedFromDocument(name) {
  test(function() { assert_equals(document[name], undefined) })
}
var documentNuked = [
  "createCDATASection",
  "createAttribute",
  "createAttributeNS",
  "inputEncoding",
  "xmlEncoding",
  "xmlStandalone",
  "xmlVersion",
  "strictErrorChecking",
  "domConfig",
  "normalizeDocument",
  "renameNode"
]
documentNuked.forEach(isNukedFromDocument)

function isNukedFromElement(name) {
  test(function() {
    var ele = document.createElementNS("test", "test")
    assert_equals(document.body[name], undefined)
    assert_equals(ele[name], undefined)
  })
}
var elementNuked = [
  "getAttributeNode",
  "getAttributeNodeNS",
  "setAttributeNode",
  "removeAttributeNode",
  "schemaTypeInfo",
  "setIdAttribute",
  "setIdAttributeNS",
  "setIdAttributeNode"
]
elementNuked.forEach(isNukedFromElement)

</script>
